+++ activesupport/lib/active_support/core_ext/object/itself.rb
@@ -1,10 +1,10 @@
 class Object
   # Returns the object itself. Useful when dealing with a chaining scenario, like Active Record scopes:
   #
-  #   Event.public_send(state.presence_in?([ :trashed, :drafted ]) ? :self).order(:created_at)
+  #   Event.public_send(state.presence_in?([ :trashed, :drafted ]) || :itself).order(:created_at)
   #   
   # @return Object
-  def self
+  def itself
     self
   end
 end